[Refactor] Create a Business controller for api endpoints in router.js
