+++ server/router.js
@@ -1,10 +1,8 @@
 const express = require('express');
 const axios = require('axios');
 const multer = require('multer');
-const { openCage, yelp, ipGeolocation, googleMapsKey } = require('./env-keys');
+const { openCage, ipGeolocation, googleMapsKey } = require('./env-keys');
 const auth = require('./middleware/auth');
-const algorithm = require('./algorithm');
-const getGFBiz = require('./getGFBiz');
 
 const router = new express.Router();
 
@@ -12,11 +10,11 @@ const router = new express.Router();
 const LocationService = require('./services/LocationService');
 const LocationServiceInstance = new LocationService();
 
-const UserService = require('./services/UserService');
-const UserServiceInstance = new UserService();
-
 const UserController = require('./controllers/UserController');
 const UserControllerInstance = new UserController();
+
+const BusinessController = require('./controllers/BusinessController');
+const BusinessControllerInstance = new BusinessController();
 
 router.get('/', (req, res) => {
   res.send('This is the root response!');
@@ -49,57 +47,10 @@ router.get('/forwardgeocode', async (req, res) => {
   }
 });
 
-const YELP_API_KEY = yelp;
 
-router.get('/yelp', async (req, res) => {
-  let lat = req.query.latitude;
-  let lng = req.query.longitude;
-  let searchbox = req.query.searchbox;
-  const options = {
-    headers: { Authorization: `Bearer ${YELP_API_KEY}` },
-  };
-  const response = await axios.get(
-    `https://api.yelp.com/v3/businesses/search?term=${searchbox}&latitude=${lat}&longitude=${lng}`,
-    options
-  );
-  let data = response.data.businesses;
-  for (let i = 0; i < data.length; i++) {
-    let newObject = {
-      name: data[i].name,
-      coordinates: data[i].coordinates,
-      location: data[i].location.display_address,
-      image: data[i].image_url,
-      price: data[i].price,
-      phone: data[i].phone,
-      categories: data[i].categories,
-      id: data[i].id,
-    };
-    data[i] = { ...newObject };
-  }
-  let coords = {
-    latitude: lat,
-    longitude: lng,
-  };
-  console.log(data);
-  data = await algorithm(data, coords);
-  res.send(data);
-});
+router.get('/yelp', BusinessControllerInstance.query);
 
-router.get('/yelp/business/:id', async (req, res) => {
-  const options = {
-    headers: { Authorization: `Bearer ${YELP_API_KEY}` },
-  };
-  const inGFDB = await getGFBiz(req.params.id);
-  if (inGFDB.length) {
-    res.send(inGFDB[0]);
-  } else {
-    const response = await axios.get(
-      `https://api.yelp.com/v3/businesses/${req.params.id}`,
-      options
-    );
-    res.send(response.data);
-  }
-});
+router.get('/yelp/business/:id', BusinessControllerInstance.getBusinessByID);
 
 /* User API Endpoints */
